+++ python/build_datasets_debug.py
@@ -78,10 +78,10 @@ base_parameters = {"env_width": env_width, "env_length": env_length, "env_height
 #env = u3_env.create_environment(task_id, base_parameters)
 
 # Linux
-env = u3_env.create_environment_by_name(file_name=f"{os.path.dirname(os.path.abspath(__file__))}/../unity/Builds/WorldDatasetGenerator/XLand", task_index=task_id, parameters=base_parameters)
+env = u3_env.create_environment(file_name=f"{os.path.dirname(os.path.abspath(__file__))}/../unity/Builds/WorldDatasetGenerator/XLand", task_index=task_id, parameters=base_parameters)
 
 # Windows
-#env = u3_env.create_environment_by_name(file_name=f"{os.path.dirname(os.path.abspath(__file__))}/../unity/Builds/WorldDatasetGeneratorWindows/unitylearning2", task_index=task_id, parameters=base_parameters)
+#env = u3_env.create_environment(file_name=f"{os.path.dirname(os.path.abspath(__file__))}/../unity/Builds/WorldDatasetGeneratorWindows/unitylearning2", task_index=task_id, parameters=base_parameters)
 
 root_folder = f"{os.path.dirname(os.path.abspath(__file__))}/../Datasets/"
 save_folder = f"{root_folder}{dataset_name}/"
